Finished Dictionary/ValidWord functionality + test
Dictionary now reads in the words and stores them in a trie and valid
words has been updated to make use of the dictionary.  Test changes
require sdk version 23 so i'm not sure if i'll keep them.  They would
allow the dictionary to be create on main screen and then later utilized
on screen 3.